WIP: - add movie presenter
