+++ src/presenter/movies-list.js
@@ -3,28 +3,28 @@ import MoviesContainerView from '../view/movies-container.js';
 import MoviesListView from '../view/movies-list.js';
 import MoviesExtraListView from '../view/movies-extra-list.js';
 import ShowMoreBtnView from '../view/show-more-btn.js';
-import MovieCardView from '../view/movie-card.js';
+// import MovieCardView from '../view/movie-card.js';
 import EmptyMovieListView from '../view/empty-movie-list.js';
 
-import MovieDetailedCardView from '../view/movie-detailed-card.js';
-import MovieCommentsView from '../view/movie-comments.js';
+// import MovieDetailedCardView from '../view/movie-detailed-card.js';
+// import MovieCommentsView from '../view/movie-comments.js';
 
 import { render, remove } from '../util/render.js';
 import { RenderPosition, FilmExtraListTitle } from '../util/const.js';
 
-// import MovieCardPresenter from './movie-card.js';
+import MoviePresenter from './movie.js';
 
 const NUMBER_OF_MOVIES_TO_RENDER = 5;
 const SECTION_MOVIES_COUNT = 2;
 const EXTRA_LIST_MOVIES_COUNT = 0;
 
-const classesToOpenDetailedFilmCard = [
-  'film-card__poster',
-  'film-card__comments',
-  'film-card__title',
-];
+// const classesToOpenDetailedFilmCard = [
+//   'film-card__poster',
+//   'film-card__comments',
+//   'film-card__title',
+// ];
 
-const siteBodyElement = document.querySelector('body');
+// const siteBodyElement = document.querySelector('body');
 
 export default class MoviesList {
   constructor(container) { //mainSiteComponent
@@ -39,8 +39,8 @@ export default class MoviesList {
 
     this._handleShowMoreBtnClick = this._handleShowMoreBtnClick.bind(this);
 
-    this._escKeyDownHandler = this._escKeyDownHandler.bind(this);
-    this._popupCloseBtnClickHandler = this._popupCloseBtnClickHandler.bind(this);
+    // this._escKeyDownHandler = this._escKeyDownHandler.bind(this);
+    // this._popupCloseBtnClickHandler = this._popupCloseBtnClickHandler.bind(this);
   }
 
   init(movies, comments) {
@@ -62,61 +62,61 @@ export default class MoviesList {
     render(this._filmsContainer, this._emptyMovieListComponent, RenderPosition.BEFOREEND);
   }
 
-  _renderDetailedFilmCardComponent() {
-    render(siteBodyElement, this._detailedFilmCardComponent, RenderPosition.BEFOREEND);
-  }
+  // _renderDetailedFilmCardComponent() {
+  //   render(siteBodyElement, this._detailedFilmCardComponent, RenderPosition.BEFOREEND);
+  // }
 
-  _renderDetailedFilmCardCommentsComponent(movie) {
-    const movieCommentsComponent = new MovieCommentsView(movie, this._comments);
-    render(this._detailedFilmCardComponent.getElement().querySelector('.film-details__bottom-container'),
-      movieCommentsComponent,
-      RenderPosition.BEFOREEND,
-    );
-  }
+  // _renderDetailedFilmCardCommentsComponent(movie) {
+  //   const movieCommentsComponent = new MovieCommentsView(movie, this._comments);
+  //   render(this._detailedFilmCardComponent.getElement().querySelector('.film-details__bottom-container'),
+  //     movieCommentsComponent,
+  //     RenderPosition.BEFOREEND,
+  //   );
+  // }
 
-  _escKeyDownHandler(evt) {
-    if (evt.key === 'Escape' || evt.key === 'Esc') {
-      evt.preventDefault();
-      remove(this._detailedFilmCardComponent); //будет undefined если вынести как метод класса если не сделать bind в конструкторе - посмотреть 4 лайв про это
-      siteBodyElement.classList.remove('hide-overflow');
-      document.removeEventListener('keydown', this._escKeyDownHandler);
-    }
-  }
+  // _escKeyDownHandler(evt) {
+  //   if (evt.key === 'Escape' || evt.key === 'Esc') {
+  //     evt.preventDefault();
+  //     remove(this._detailedFilmCardComponent); //будет undefined если вынести как метод класса если не сделать bind в конструкторе - посмотреть 4 лайв про это
+  //     siteBodyElement.classList.remove('hide-overflow');
+  //     document.removeEventListener('keydown', this._escKeyDownHandler);
+  //   }
+  // }
 
-  _popupCloseBtnClickHandler() {
-    remove(this._detailedFilmCardComponent);
-    siteBodyElement.classList.remove('hide-overflow');
-    document.removeEventListener('keydown', this._escKeyDownHandler);
-  }
+  // _popupCloseBtnClickHandler() {
+  //   remove(this._detailedFilmCardComponent);
+  //   siteBodyElement.classList.remove('hide-overflow');
+  //   document.removeEventListener('keydown', this._escKeyDownHandler);
+  // }
 
-  _renderDetailedFilmCard(movie, evt) {
-    const clickTarget = evt.target.classList.value;
+  // _renderDetailedFilmCard(movie, evt) {
+  //   const clickTarget = evt.target.classList.value;
 
-    if (classesToOpenDetailedFilmCard.includes(clickTarget)) {
-      siteBodyElement.classList.add('hide-overflow');
-      this._detailedFilmCardComponent = new MovieDetailedCardView(movie);
+  //   if (classesToOpenDetailedFilmCard.includes(clickTarget)) {
+  //     siteBodyElement.classList.add('hide-overflow');
+  //     this._detailedFilmCardComponent = new MovieDetailedCardView(movie);
 
-      if (!siteBodyElement.querySelector('.film-details')) {
-        this._renderDetailedFilmCardComponent();
-        this._renderDetailedFilmCardCommentsComponent(movie);
+  //     if (!siteBodyElement.querySelector('.film-details')) {
+  //       this._renderDetailedFilmCardComponent();
+  //       this._renderDetailedFilmCardCommentsComponent(movie);
 
 
-        this._detailedFilmCardComponent.setCloseBtnClickHandler(this._popupCloseBtnClickHandler);
-        document.addEventListener('keydown', this._escKeyDownHandler);
-      }//проверяет есть ли уже карточка
-    }//проверяет был ли клик по тому элементу, который открывает попап
-  }//renderDetailedFilmCard
+  //       this._detailedFilmCardComponent.setCloseBtnClickHandler(this._popupCloseBtnClickHandler);
+  //       document.addEventListener('keydown', this._escKeyDownHandler);
+  //     }//проверяет есть ли уже карточка
+  //   }//проверяет был ли клик по тому элементу, который открывает попап
+  // }//renderDetailedFilmCard
 
   _renderFilmCard(filmCardcontainer, movie) {
-    const filmComponent = new MovieCardView(movie);
-    render(filmCardcontainer, filmComponent, RenderPosition.BEFOREEND);
+    // const filmComponent = new MovieCardView(movie);
+    // render(filmCardcontainer, filmComponent, RenderPosition.BEFOREEND);
 
-    filmComponent.setOpenDetailedFilmCardHandler((evt) => {
-      this._renderDetailedFilmCard(movie, evt);
-      // const filmDetailedCardPresenter = new MovieCardPresenter(movie, this._comments);
-      // filmDetailedCardPresenter.init(movie, evt);
+    // filmComponent.setOpenDetailedFilmCardHandler((evt) => {
+    // this._renderDetailedFilmCard(movie, evt);
+    const filmCardPresenter = new MoviePresenter(filmCardcontainer, this._comments);
+    filmCardPresenter.init(movie);
 
-    });
+    // });
   }
 
   _renderFilmListContainer() {
